# Implement fastq_iter as iterator class