… src/dnaio/_core.pyi