Fixes #31 

The Python C-API does not know the `yield` keyword. What it does know are classes with `tp_iter` and `tp_iternext` set. By implementing `fastq_iter` as a class it resembles the Python C-API more closely. Also it allows us to track the number of records that have been yielded so far and make it an accessible attribute. This has advantages for paired-end reading where we can use `zip` rather than `itertools.zip_longest`.

Before:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):     989.9 ms ±  20.6 ms    [User: 873.8 ms, System: 116.0 ms]
  Range (min … max):   962.9 ms … 1030.4 ms    10 runs
 
Benchmark #1: python dnaio_read_paired.py ~/test/big2.fastq
  Time (mean ± σ):      2.679 s ±  0.023 s    [User: 2.478 s, System: 0.201 s]
  Range (min … max):    2.651 s …  2.717 s    10 runs
```
+ Speed single end: 990 ms
+ Speed paired end: 2680 ms
+ Overhead paired end: 2680 - (990*2) = 700ms. (26%)

After:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):     909.5 ms ±  19.2 ms    [User: 786.6 ms, System: 122.8 ms]
  Range (min … max):   886.1 ms … 955.9 ms    10 runs
 
/big2.fastq'
Benchmark #1: python dnaio_read_paired.py ~/test/big2.fastq
  Time (mean ± σ):      2.366 s ±  0.019 s    [User: 2.171 s, System: 0.195 s]
  Range (min … max):    2.342 s …  2.399 s    10 runs

```
+ Speed single end: 910 ms
+ Speed paired end: 2370 ms
+ Overhead paired end: 2370 - (910*2) = 550ms. (23%)

The overhead reduction for paired-end is not much, but that is because the name checking is a big part of that one. It is quite nice that this improves the read speed for single end data even further, and the pairedend reading also reaps the benefit from that.

@marcelm, even though the diff looks large, this is basically the same code as before in fastq_iter. Now there is more usage of self, but that is about it (the changes to make it work in an iterator were quite minor). The changes are in the python code area.