@rhpvorderman 2022-02-07
The speed difference in single-end can be explained by how Cython handles the code. With this code it does simply as told. It creates an iterator class. With the function with the `yield` keyword it defines code for a custom Cython CoRoutine class as well as making fastq_iter a class. I can't exactly grok what Cython is doing, but it seems clear that quite a lot of performance is lost in the translation. Probably due to generators having a return type and that they can allow signals to be sent to it. This has to be implemented in C as well, so that will create some overhead. 

The generated C code for this PR is 900 lines shorter than for the code in main as a result.