@marcelm 2022-02-07
Thanks, I’ll have a look later this week. Great that this improves speed even further. By the way, what is the size of your `big2.fastq` file? (I’d like to know what the throughput is in GB/s.)