@rhpvorderman 2022-02-07
It is exactly 5 million FASTQ records Illumina HiSeq 4000 data. (150 bp per read, 336 bytes per record). The total filesize is `1675000000` bytes.