@rhpvorderman 2022-02-11
The great improvement is probably due to having quite a lot of records. 43 bp is not much, so the record-to-byte ratio is quite high. It seems that cython's implementation of `yield` was quite inefficient compared to `__next__` (which is a normal function call). 
The improvement will be less pronounced on longer reads.

>I’ll have to go now – sorry I cannot look at the changes you made just now; I may have time over the weekend.

Take your time!