@marcelm 2022-02-14
You are right regarding the overhead for short reads. Here are some measurements on SRR2174302 (randomly picked PacBio file from SRA, has an average read length of 9025 bp), extracted with `fastq-dump --defline-qual '+' --defline-seq '@$ac.$si' SRR2174302`. Using Python 3.8.

|tag|ns/byte|
|-|-|
|v0.5.0|0.719|
|v0.6.0|0.718|
|v0.7.1|0.220|
|9246fa9d|0.188|
|this PR|0.188|

It’s superfast, but not because of this PR in particular :smiley:.

Everything looks good, will merge now!